requirements.txt
google-api-python-client
python-dotenv
pandas
isodate
scikit-learn
streamlit
plotly